Implement jwt for user controller also allow admins to delete all accounts, but users only to delete own accounts

+++ backend-blog-api/controllers/userController.js
@@ -1,16 +1,22 @@
 const { prisma } = require("../lib/prisma.js")
 
 async function deleteUser(req, res) {
+    const targetUserId = Number(req.params.userId);
+    const requestingUserId = req.user.userId;
+    const requestingUserRole = req.user.role;
     try {
         const userExists = await prisma.user.findUnique({
             where: {
-                id: Number(req.params.userId)
+                id: targetUserId
             },
         });
-        if (!userExists) return res.status(404).json({error: "user not found"});
+        if (!userExists) return res.status(404).json({ error: "user not found" });
+
+        if (requestingUserRole.role !== "ADMIN" && requestingUserId !== targetUserId) return res.status(403).json({ error: "you may only delete your own account" });
+
         const deletedUser = await prisma.user.delete({
             where: {
-                id: Number(req.params.userId)
+                id: targetUserId
             },
             select: {
                 id: true,
@@ -18,11 +24,10 @@ async function deleteUser(req, res) {
                 email: true
             }
         });
-
         res.json({ deletedUser: deletedUser })
     } catch (err) {
         console.error(err);
-        res.status(500).error({ error: "error deleting user" })
+        res.status(500).json({ error: "error deleting user" })
     }
 }
